+++ Common/src/main/java/net/potionstudios/woodwevegot/world/level/block/WWGTrappedChestBlock.java
@@ -10,13 +10,13 @@ import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.block.entity.ChestBlockEntity;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.state.BlockState;
-import net.potionstudios.woodwevegot.world.level.block.entity.WWGBlockEntities;
+import net.potionstudios.woodwevegot.world.level.block.entity.WWGBlockEntityType;
 import org.jetbrains.annotations.NotNull;
 
 public class WWGTrappedChestBlock extends WWGChestBlock {
 
     protected WWGTrappedChestBlock(BlockBehaviour.Properties properties, String set) {
-        super(() -> WWGBlockEntities.TRAPPED_CHEST.get(), properties, set);
+        super(() -> WWGBlockEntityType.TRAPPED_CHEST.get(), properties, set);
     }
 
     @Override